==> notebooks/13_Motif_disruprion_analysis_prepare_table.ipynb <==
## Prepare the tables necessary to calculate motif disruption
1. Get the SNPs to test: in this case I am testing all heterozygous SNPs that were tested for ASE in ChIP-Seq
2. Get the fasta for a 41 bp window (20 bases upstream and downstream), for both alleles
3. Get the results of the 20 most enrichemt motif e from Homer
4. Run the analysis using the script scoreMotifs_m1.R (computationally intense and long)

library (stringr)
home="/home/paola/Family1070/private_output/"
dir="ASE_chip/correlationWithBeta/"
names=c("NKX25", "H3K27AC_IPSC",  "H3K27AC_CM","ATAC_IPSC", "ATAC_CM")  #also names of directories
files= paste('ASE.tested', names, 'txt', sep=".")
setwd(home)
setwd("Motif_analysis_R")

FILES = files[1]
name  = names[1]
#dir.create(name)

ase=read.table(paste(home,dir,FILES,sep=""), sep="\t", header=T)
test=str_split_fixed(ase$variantID, ":", 2)
ase$chr=test[,1]
ase$POS=as.numeric(test[,2])
ase$bed.start=ase$POS-20
ase$bed.end=ase$POS+20

write.table(ase[,c("chr", "bed.start", "bed.end","variantID")],
            paste(name, "all_snps.bed", sep="/"), row.names=F, col.names=F, sep="\t", quote=FALSE)



command= paste("bedtools getfasta -fi /publicdata/hg19_20151104/hg19_sorted.fa -bed  ", 
               name, "/all_snps.bed -fo " , name,  "/all_fasta.tab -name -tab", sep="")

system("module load cardips")
system(command)


  file=paste(name, "all_fasta.tab", sep="/")
  fas=read.table(file)
  
  fas=cbind(fas,ase[,c("refAllele", "altAllele", 'ref_freq', 'combined_fdr')])
  # fas=cbind(fas,ase[,c("refAllele", "altAllele", 'new_maf', 'combined_fdr')]) # colum name of previous file
  colnames(fas)=c("pos","seq","ref","alt",'ref_freq', 'combined_fdr')
  fas[,1:4]=sapply(fas[,1:4], as.character)
  fas1=fas
  
  
  # Commands for writing fasta format
  Xfasta <- character(nrow(fas) * 2)
  Xfasta[c(TRUE, FALSE)] <- paste0(">", fas$pos)
  Xfasta[c(FALSE, TRUE)] <- fas$seq
  writeLines(Xfasta, paste(name,"ref.fasta", sep="/"))

  ###########
  for (i in 1:nrow(fas1)){
   
      substr(fas1[i,"seq"], 20, 20) <- fas1[i,"alt"]
    
  }
  
  # Commands for writing fasta format
  Xfasta <- character(nrow(fas1) * 2)
  Xfasta[c(TRUE, FALSE)] <- paste0(">", fas1$pos)
  Xfasta[c(FALSE, TRUE)] <- fas1$seq
  writeLines(Xfasta, paste(name,"alt.fasta", sep="/"))
  


##### Make table to use in Erin's program
matr=cbind(fas, fas1[,c("seq")])
matr=matr[c(1,3,4,5,6,2,7)]
#adapt column names to erin's script 
colnames(matr)[c(1,6,7)]=c("Variant", "REfSeq", "ALtSeq")
matr$REfSeq<-toupper(matr$REfSeq)
matr$ALtSeq<-toupper(matr$ALtSeq)

write.table(matr, paste(name, "TestMAtr.txt", sep="/"), row.names=F, col.names=T, sep="\t", quote=F)



head(matr)

motif_dir_new = paste(name, "analysis","PWMs", sep="/")
dir.create(motif_dir_new)

motif_dir = paste(home, "Homer_motif_enrichment/", name, "/knownResults/", sep="")

homer_scores = data.frame()
for( i in 1:20){
   
   pwm = read.table(paste(motif_dir, "known",i, ".motif", sep=""), sep="\t", skip=1)
   colnames(pwm)=c('A','C','G','T')
   
    head = read.table(paste(motif_dir, "known",i, ".motif", sep=""), sep="\t", nrows=1, stringsAsFactor=F) 
    homer_scores[i,1]= str_split_fixed(head$V2, "\\(", 2 )[,1]
    homer_scores[i,2]= head$V3
    
    write.table(pwm, paste(motif_dir_new, homer_scores[i,1], sep="/" ),  sep="\t",  row.names=F, quote=F) 
}


write.table(homer_scores, paste(name, "analysis", "HomerScores.txt" , sep="/"), quote= F, col.names=F, row.names=F) 

To run the analysis, qsub the code 2_scoreMotifs_m1.R  --- it takes about an hour per motif.
