==> notebooks/17_GWAS-annotations.ipynb <==
### Annotation of SNPs with GWAS hits and LD SNPs
1. GWAS SNPS obtained from the GWAS catalog on (April 17/2017) plus manuall addition of the three new atrial fibrillation studies from April 2017-----> UPDATED CATALOG DOWNOLOADED ON 10-20-2017 ---IMPORTANT NOTE DOWNLOAD THE VERSION V.1.0.1 BECAUSE THERE IS A UNIFORM MAPPED_TRAIT COLUMN
2. Using the tables from Haploreg (1000G phase I)get all SNPs in LD
3. Manual selection of EKG traits

#setwd('/home/paola/2017/anotations/GWAS')
setwd('/home/paola/Family1070/private_output/Enrichment_annotations/GWAS')
library (stringr)

#gwas<-read.csv('gwas_catalog_2017-04-24_mod.csv')
gwas<-read.csv('gwas_catalog_v1.0.1-associations_e90_r2017-10-10_mod.txt', header=T, sep="\t")
trait_list<-read.csv("TraitList.csv" )

dim(gwas)
gwas=subset(gwas, gwas$SNP_ID_CURRENT!="rs")
dim(gwas)

writeLines(as.character(gwas$SNP_ID_CURRENT), 'GwasCatSnps.txt')

gwas<-merge(gwas, trait_list, by="MAPPED_TRAIT", all.x=T)

system("zcat /publicdata/haploreg_v4.0_20151021/LD_ASN.tsv.gz | awk 'NR==FNR{a[$0]=$0}NR>FNR{if($1==a[$1])print $0}' GwasCatSnps.txt - >  GwasCatSnps.LD_ASN.txt")

system("zcat /publicdata/haploreg_v4.0_20151021/LD_EUR.tsv.gz | awk 'NR==FNR{a[$0]=$0}NR>FNR{if($1==a[$1])print $0}' GwasCatSnps.txt - >  GwasCatSnps.LD_EUR.txt")

# tab=c("GwasCatSnps.LD_ASN.txt", "GwasCatSnps.LD_EUR.txt")
 
#   for(i in 1:2){
#     ld<-read.table(tab[i])
    
#     colnames(ld)<-c("lead","LD")
#     x<-sapply(ld$LD, str_count, ";")
    
#     test=str_split_fixed(ld$LD, ";", max(x))
    
#     #preallocate matrix is faster
    
#     nr=(dim(test)[1]*dim(test)[2])-sum(test=="")
#     newmat=matrix(NA, nrow=nr, ncol=2)
#     k=1
#     for (a in 1:nrow(test)){
#       for(j in 1:ncol(test)){
#         if(test[a,j]!="") {
#           newmat[k,1]<-as.vector(ld$lead)[a]
#           newmat[k,2]<-test[a,j]
#           k=k+1
#         }
#       }
#     }
  
#     tfx<-data.frame(str_split_fixed(newmat[,2], "\\,",3 ))
#     colnames(tfx)<-c("LDsnp", "r2", "D")
#     tfx$leadSNP<-newmat[,1]
#     tfx<-tfx[!duplicated(tfx),]
#     write.table(tfx, paste("Melted",tab[i], sep="."), sep="\t", col.names=T, row.names=F, quote=F)
   
#   }


eur<-read.table("Melted.GwasCatSnps.LD_EUR.txt", header=T)
eur<-subset(eur, eur$r2>0.85)
asn<-read.table("Melted.GwasCatSnps.LD_ASN.txt", header=T)
asn<-subset(asn, asn$r2>0.85)

eur<-merge(gwas, eur[,c(1,2,4)], by.x="SNP_ID_CURRENT", by.y="leadSNP", all.x=T)
asn<-merge(gwas, asn[,c(1,2,4)], by.x="SNP_ID_CURRENT", by.y="leadSNP", all.x=T)
asn<-asn[!duplicated(asn),]
eur<-eur[!duplicated(eur),]


write.csv(asn,"Table.GwasCatSnps.08LD_ASN.csv", row.names=F )
write.csv(eur,"Table.GwasCatSnps.08LD_EUR.csv", row.names=F )


head(eur)

unique(eur$MAPPED_TRAIT[eur$Electro==T])

###########################################################
### Shape Table for final annotation table
###########################################################
### Merge gwas pubmedID
gwas<-gwas[!duplicated(gwas),]
stud<-subset(gwas, !duplicated(gwas[, c("MAPPED_TRAIT","SNP_ID_CURRENT", "PUBMEDID")]))
stud$gwas<-paste(stud$SNP_ID_CURRENT, stud$MAPPED_TRAIT, sep="|")
agg1<-aggregate(PUBMEDID~gwas, stud, paste, collapse=",")
agg1$study<-paste(agg1$gwas,agg1$PUBMEDID, sep="|")
stud<-merge(agg1, stud, by="gwas")
stud<-stud[,c("SNP_ID_CURRENT", "study")]


## Merge LD population frequency
asn$pop<-"ASN"
eur$pop<-"EUR"
all<-rbind(asn, eur)
all$r2_pop<-paste(all$r2, all$pop, sep="_")
all<-subset(all, !is.na(all$LDsnp))
all<-subset(all, !duplicated(all[,c('SNP_ID_CURRENT', 'LDsnp', "r2_pop")]))
all$SNP<-paste(all$SNP_ID_CURRENT, all$LDsnp, sep="|")
agg2<-aggregate(r2_pop~SNP,all, paste, collapse=",")
ld<-merge(agg2, all, by="SNP")
ld<-ld[,c('r2_pop.x', 'SNP_ID_CURRENT', 'LDsnp')]
ld<-ld[!duplicated(ld),]

### Merge all

m<-merge(stud, ld, by="SNP_ID_CURRENT", all=T)

write.csv(m,"GwasForFinalAnnotation.csv", row.names=F)

head(m)
