==> notebooks/20_Supplementary_table_annotation.ipynb <==
# Make final supplementary table

library (stringr)
home = "/home/paola/Family1070/private_output/"
setwd(paste(home,"Enrichment_annotations", sep="/"))
ase_dir="/home/paola/Family1070/private_output/ASE_chip/pipeline/results"
peaks = list.files("/frazer01/home/paola/Family1070/private_output/PeakCalling/")


marks= peaks[3:5]

###### Closest Gene Annotation
gi   = read.table("/frazer01/publicdata/gencode_v19_20151104/gene_info.tsv", header=T)
pro  = paste(home, "Coordination_ASE_effects/matches/", marks, "_closest_gene.bed", sep="" )

##### Motif disruption calculated here 

motR = '/frazer01/home/paola/Family1070/private_output/Motif_analysis_R/NKX25/analysis/AlleleAlteringSites_scores_1.txt'
motR2 = '/frazer01/home/paola/Family1070/private_output/Motif_analysis_R/NKX25/analysis_denovo/AlleleAlteringSites_scores_1.txt'
motR3 = '/frazer01/home/paola/Family1070/private_output/Motif_analysis_R/NKX25/analysis_denovo_part2/AlleleAlteringSites_scores_1.txt'
        
### Haploreg annotations

HAP=paste ( "Haploreg/", marks, "_haploreg.tab", sep="")
QTL=paste ( "GTEx/", marks, "_eQTLs.txt", sep="")

#### Gwas annotations
gwasCAT<- read.csv('GWAS/GwasForFinalAnnotation.csv', header=T)

  gwasCAT$anno<-paste(gwasCAT$study,gwasCAT$LDsnp,gwasCAT$r2_pop.x, sep='|')
  lead<-subset(gwasCAT, !duplicated(gwasCAT[,c("SNP_ID_CURRENT", "study")]))
  lead<-aggregate(study~SNP_ID_CURRENT, lead, paste, collapse=";")
  ld<-subset(gwasCAT, !duplicated(gwasCAT[,c("LDsnp", "anno")]))
  ld<-aggregate(anno~LDsnp, ld, paste, collapse=";")

##### left ventricle expression annot
lv<-read.csv('GTEx/GTEx Portal.csv', header=T)
aa<-read.csv('GTEx/GTEx Portal AA.csv', header=T)

lv_eqtls<-read.table("/publicdata/gtex_v6/Heart_Left_Ventricle_Analysis.snpgenes",  sep="\t", header=T)
lv_eqtls$snpID<-paste("chr",lv_eqtls$snp_chrom, ":", lv_eqtls$snp_pos, sep="")
lv<-merge(lv, lv_eqtls[,c('gene', 'snpID' )], by.x='Gencode.Id', by.y='gene')

aa_eqtls<-read.table("/publicdata/gtex_v6/Heart_Atrial_Appendage_Analysis.snpgenes",  sep="\t", header=T)
aa_eqtls$snpID<-paste("chr",aa_eqtls$snp_chrom, ":", aa_eqtls$snp_pos, sep="")
aa<-merge(aa, aa_eqtls[,c('gene', 'snpID' )], by.x='Gencode.Id', by.y='gene')



m=data.frame()

for (i in 1:length(marks)){
  
  cat(marks[i], "\n")
  mark=marks[i]
  ase=read.csv(paste("Haploreg/SummaryAnnotations", mark, "csv", sep="."), header=T)
  ase=subset(ase, ase$combined_fdr<0.05)
  
  
  hap=read.table(HAP[i], header=F, sep="\t")
  colnames(hap)=c("chrom", "pos", "id", "ref", "alt", "cons", "motif", "gene","chromHMM", "DNAse")
  hap=hap[,c( "id", "motif", "gene","chromHMM", "DNAse")]
  ase=merge(ase,hap,by.x="snpID", by.y="id", all.x=T)
 
  qtl<-read.table( QTL[i], fill=T, stringsAsFactors=F)
  colnames(qtl)<-c("snp","qtl")
  ase=merge(ase,qtl,by.x="snpID", by.y="snp", all.x=T)
  
  gene=read.table(pro[i])
  colnames(gene)=c("chr", "start", "end", "chr2", 
                   "start2", "end2", "gene", "dot", "strand", "dist")
  gene$locus = paste (gene$chr,":", gene$start, "-" ,gene$end, sep="")
  gene=merge(gene, gi, by.x="gene", by.y="gene_id")
  gene$closestGenes=paste(gene$gene, gene$gene_name, gene$dist, sep="|")
  gene=gene[,c("locus", "closestGenes")]
  gene=aggregate(closestGenes~locus, gene, paste, collapse=",")
  ase=merge(ase,gene,by="locus", all.x=T)
  
  
  ase<-merge(ase, lead[, c("SNP_ID_CURRENT", "study")], by.x="snpID", by.y="SNP_ID_CURRENT", all.x=T)
  ase<-merge(ase, ld[, c("LDsnp", "anno")], by.x="snpID", by.y="LDsnp", all.x=T)
  
eqtls = read.table( paste("GTEx/", mark,  ".GtexTable_hreg_aggregate.txt", sep=""), header=T,  sep="\t") 
heart_specific = subset(eqtls, ((eqtls[,2]/2)-eqtls[,"Hear"])<=0)
ase$heart_specific = ase$snpID %in% heart_specific$snpID
  
    if (mark == "NKX25"){
  mymot  = read.table(motR, header=T,stringsAsFactor=F)
  mymot2 = read.table(motR2, header=T,stringsAsFactor=F)
  mymot3 = read.table(motR3, header=T,stringsAsFactor=F)   

  mymot=rbind(mymot,mymot2,mymot3)
  mymot[is.na(mymot)]<-0
  mymot=merge(mymot, ase[,c('variantID', 'ref_freq')], by.x="SNP", by.y="variantID")

  mymot$Diff<-round(mymot$maxAltScore-mymot$maxRefScore,2)  
  mymot$Motif_anno<-paste(mymot$Motif, mymot$Diff,mymot$maxRefScoreStrand, 
                          round(mymot$maxRefScore,2), round(mymot$maxAltScore,2), sep="|" )
  agg=aggregate(Motif_anno~SNP, mymot, paste, collapse=",")
  ase=merge(ase, agg, by.x="variantID", by.y="SNP", all.x=T)
 
  }
  
  
  else{
    
    ase$Motif_anno<-NA
    
  }
    

 if(i>1)      {ase<-subset(ase, select=colnames(m))   }
  
    
    m<-rbind(m,ase)
  
}


m$topxp<-m$snpID %in% lv$snpID | m$snpID %in% aa$snpID
m$bestp<-m$snpID %in% lv_eqtls$snpID[lv_eqtls$has_best_p==1] | 
m$snpID %in% aa_eqtls$snpID[aa_eqtls$has_best_p==1]


write.table(m, "Supplementary_table_ase_annotations.txt", sep="\t", row.names=FALSE, quote=FALSE)
