==> notebooks/22_Validate_SNPs_in_CARDiPS_prepare_tables.ipynb <==
## Prepare tables of gene expression and genotypes to use for associations

home = "/frazer01/home/paola/Family1070/private_output/Validation/data_tables"
setwd(home)


suppressPackageStartupMessages(library(stringr))
suppressPackageStartupMessages(library(DESeq2))
suppressPackageStartupMessages(library(limma))

### 1. Get all rna-seq uuid done so far in iPSC-CMs 
this table (from Margaret D.) contains all cardips uuis for different assays. Only succesfull iPSC-CMs (cTNT>35%) were included


main = read.table ("/frazer01/projects/CARDIPS/data/CARDiPS_production_metadata/assay_mastertable.txt", sep= "\t", header=T)
main= main[!duplicated(main),]

table(main$assay, main$dataset)

rna_cm = droplevels(subset(main, dataset=="production" &assay =="rna" & cell=="CM"))
dim(rna_cm)

head(rna_cm)

### 2. Get the gene expression RNA-Seq data 
normalize, correct for TNNT and export it

raw<-read.table("/projects/CARDIPS/pipeline/RNAseq/combined_files/rsem_expected_counts.tsv", header=T, check.names=F, 
                  sep="\t", row.names=1)

tpm<-read.table("/projects/CARDIPS/pipeline/RNAseq/combined_files/rsem_tpm.tsv", header=T, check.names=F, 
                  sep="\t", row.names=1)

gene_info<-read.table("/publicdata/gencode_v19_20151104/gene_info.tsv", header=T, sep="\t")

chromo<-str_split_fixed(gene_info$chrom, "chr", 2)[,2]
unique(chromo)
autoso<- as.character(1:22)


raw_sel<-raw[chromo %in% autoso,]
tpm_sel<-tpm[chromo %in% autoso,]

raw_sel<-subset(raw_sel, select=c(as.character(rna_cm$assay_uuid)))
tpm_sel<-subset(tpm_sel, select=c(as.character(rna_cm$assay_uuid)))
raw_sel<-raw_sel[(rowMeans(tpm_sel)>2),]
tpm_sel<-tpm_sel[(rowMeans(tpm_sel)>2),]

dim(tpm_sel)
dim(raw_sel)

## put id first for Deseq
rna_cm<-rna_cm[,c(6:length(rna_cm), 1:5)]
rna_cm<-rna_cm[as.character(rna_cm$assay_uuid) %in% colnames(raw),]
ctnt<- tpm_sel[as.character(gene_info[gene_info$gene_name=="TNNT2","gene_id"]),]
ctnt<- subset(ctnt, select=as.character(rna_cm$assay_uuid))
rna_cm$cTNT_TPM = t(ctnt)[,1]

write.csv(rna_cm, "Production_cm_samples.csv")

counts<-round(raw_sel,0)
workingData<-DESeqDataSetFromMatrix(counts, rna_cm, design= ~ subject_uuid)
workingData<-estimateSizeFactors(workingData)
workingData<-estimateDispersions(workingData, fitType="parametric")
vst_workingData<-varianceStabilizingTransformation(workingData, blind=TRUE)

mat<-as.data.frame(assay(vst_workingData))
colnames(mat)<-as.character(rna_cm$assay_uuid)
write.table(mat, file="Production_cms_vst_counts.txt", quote=F, row.names=T, col.names=T, sep="\t")


rna_cm =read.csv( "Production_cm_samples.csv", row.names=1)


head(rna_cm)

mat = read.table( "Production_cms_vst_counts.txt", header=T, row.names=1, check.names=F)

### normalize using residuals from TTNT2
fit <- lmFit( mat, model.matrix(~ cTNT_TPM, rna_cm))
res <- residuals(fit, mat)
mat_corrected<- res + rowMeans(as.matrix(mat))
write.table(mat_corrected, file="Production_cms_vst_counts_corrected_ctnt.txt", quote=F, row.names=T, col.names=T, sep="\t")

### 4. Get genotypes for selected SNPs and individuals 

system("module load cardips")

snps = read.table("../../Enrichment_annotations/Supplementary_table_ase_annotations.txt",  header=T, fill=T)

nkx_snps = subset(snps, mark=="NKX25")
dim(nkx_snps)
nkx_snps = str_split_fixed(nkx_snps$variantID, ":",2)


write.table(nkx_snps, "NKX_ase_coord.txt", sep="\t", quote=F, col.names=F, row.names=F)

get_genotypes<- "bcftools query -R NKX_ase_coord.txt -f '%CHROM\t%POS\t%ID\t%REF\t%ALT[\t%GT]\n' -H /projects/CARDIPS/pipeline/WGS/mergedVCF/CARDIPS_201512.PASS.274.hg19.vcf.gz -o Cardips_NKX25_ase_gt.tsv"

system (get_genotypes)

### manually remove the # in the header

gt<-read.table("Cardips_NKX25_ase_gt.tsv", header=T,check.names=F, stringsAsFactors=FALSE )

gt <- replace(gt, gt == "1/1", 2)
gt <- replace(gt, gt == "0/1", 1)
gt <- replace(gt, gt == "0/0", 0)


names<-str_split_fixed(colnames(gt), "]", 2)[,2]
names2<-str_split_fixed(names[-c(1:5)], ":",2)[,1]
colnames(gt)<-c(names[1:5], names2)
write.csv(gt, "Cardips_NKX25_ase_genotypes_coded.csv")

dim(rna_cm)
wgs     = read.table ("wgs.txt", header=T)
rna_cm<-merge(rna_cm, wgs, by.x="subject_uuid", by.y="subject_id")
dim(rna_cm)
rna_cm=  rna_cm [!duplicated(rna_cm),]
rna_cm = subset(rna_cm, wgs_id %in% colnames(gt))
dim(rna_cm)

write.csv(rna_cm, "Production_cm_samples_wgsid.csv")
