==> notebooks/26_FGWAS_PRinterval-vanSetten.ipynb <==
### Prepare tables to run fGWAS fine mapping using chromatin annotations using Bill's pipeline

suppressPackageStartupMessages(library("tidyr"))
options(scipen=999)

suppressPackageStartupMessages(library("stringr"))

home="/home/paola/Family1070/private_output/fgwas_analysis/PR_interval/preprocessing/"
setwd(home)

gwas = read.table("/publicdata/gwas_summary_stats_20180124/vanSetten_2018/PR_interval_July2018_summary_results.hg19.txt", 
                  header=T, sep="\t")

head(gwas)

colnames(gwas)= c("SNPID", "CHR", "POS_hg18", "A1", "A2", "F", "Beta", "SE", "P_VALUE", "POS")

required = c("SNPID", "CHR", "POS", "SE", "Z", "F", "N")

gwas$Z = gwas$Beta /gwas$SE
gwas$N = '.'

gwas$POS_1 = gwas$POS -1

sum(is.na(gwas$SNPID))

Annotate with regulatory regions and ASE

write.table(gwas[,c("CHR","POS_1","POS","SNPID")], "PR_Interval.bed", row.names=F, col.names=F, quote=F, sep="\t")

annotate_gwas = function( regiofile, annofile ){
    system(paste("bedtools intersect -a" ,  regiofile, "-b", annofile , "-wo > intesect_results"))
    annot = read.table("intesect_results")  
    annot = annot[!duplicated(annot),]
    annot_wide <- spread(annot[,c(1:4,8,9)], V8, V9)
    annot_wide [is.na(annot_wide)]<-0
    colnames(annot_wide)[1:4] = c("CHR", "START", "POS", "SNPID")
    return(annot_wide)
   }

anno = annotate_gwas('PR_Interval.bed', 'frazer_lab_cm_annotations.bed')

head(anno)

m = merge(gwas, anno, by= c('CHR','POS', 'SNPID'), all.x=TRUE)

dim(anno)

m [is.na(m)]<-0

dim(m)

nkx_ase = readLines("NKX25_ASE_rsids.txt")
h3k_ase = readLines("H3K27AC_ASE_rsids.txt")

m$NKX25_ASE = as.numeric(m$SNPID %in% nkx_ase)
m$H3K27AC_ASE = as.numeric(m$SNPID %in% h3k_ase)

m = m[!duplicated(m),]

sum(duplicated(gwas$POS))

remove duplicated SNPs (merged / renames SNPs)

m = m[with(m, order(SE)),]
m = m[!duplicated( m[,c("CHR", "POS")]),]     
m = m[with(m, order(CHR, POS)),]

remove chr Y and M

m =subset(m, !(CHR %in% c("M", "Y")))

mori = subset(m, select = as.character(c("SNPID", "CHR", "POS", "P_VALUE" )))

write.table(mori, "../Original_input", col.names=T, row.names=F, sep="\t",quote=F)

mano = subset(m, select = c( required, 'ATAC','H3K27AC','NKX25','NKX25_ASE','H3K27AC_ASE' ))

write.table(mano, "Annotated_input", col.names=T, row.names=F, sep=" ",quote=F)

Remember to run the fgwas command to determine the size of the chunks for the analysis (-k)
for this analysis fgwas k = 910 ( 5,000/5.5 Mb)

### Read Results

getwd()

res = read.table('../pipeline_out/Final_Model/Final_Model.bfs', header=T)
seg = read.table('../pipeline_out/Final_Model/Final_Model.segbfs', header=T)
res = merge(res,seg[,c('chunk', 'PPA')], by="chunk" )
res$PPA = res$PPA.x *res$PPA.y
res = res[order(res$PPA, decreasing=T),]

nkx = subset(res, NKX25_ASE==1 )

head(nkx,15)

There are several complelling example for novel sub-threshold loci that could be mediated by NKX2-5, the most intereting:
- rs13006682/ FHL2 : the gene has its highest expression in heart ventricle (GTEx) and linked to cardiac hyperthophy (KO mouce no phenotype though) - the SNP distrupt a tbx motif
- rs2267469/ PACSIN2 the gene colocalize with CAV1 in the caveolae, heart weight phenotype in the mice invenotry - the snp disrupt nkx2-5 and loops directly to the prmpter of PACSIN2

** Check the new snp/gene associations in CARDIPS **

There is no eQTL for these genes.
