==> notebooks/30_LD_HiC_fGWAS_plots.ipynb <==
# Make regional plot for SNPs fine mapping





recombination rate and 1KG pilot LD data are taken from the "SNAP" web-application and copied manually to the cluster
http://archive.broadinstitute.org/mpg/snap/ldplot.php

#gwas  = read.table("/publicdata/gwas_summary_stats_20180124/denHoed/rsid_pval.txt", header=T, sep="\t")

gwas  = read.table("/publicdata/gwas_summary_stats_20180124/vanSetten_2018/PR_interval_July2018_summary_results.hg19.txt", header=T, sep="\t")

#fgwas = read.table('/home/bill/paola-heart-finemapping/Data/Optimize_Model_ASE/No_Marks_Removed/No_Marks_Removed.bfs', header=T)

#chunks = read.table('/home/bill/paola-heart-finemapping/Data/Optimize_Model_ASE/No_Marks_Removed/No_Marks_Removed.segbfs', header=T)

fgwas = read.table('~/Family1070/private_output/fgwas_analysis/PR_interval/pipeline_out/Final_Model/Final_Model.bfs', header=T)
chunks = read.table('~/Family1070/private_output/fgwas_analysis/PR_interval/pipeline_out/Final_Model/Final_Model.segbfs', header=T)


fgwas  = merge (gwas, fgwas, by.x="rsid", by.y="id")

backup = fgwas

fgwas= backup

head(fgwas)

fgwas[fgwas$rsid=="rs176107",]

log10(0.0001682)

#### produce a bed file to be lifted over from hg18 to hg19

liftover = function(snapfile){

data.file <- read.table(snapfile, header=T, na.strings="N/A")
data.file$end<-data.file$Coordinate_HG18+1
write.table(data.file[complete.cases(data.file[,c(5,6,8,2)]),c(5,6,8,2)], "data.file.bed", sep="\t", quote=F, col.names=F, row.names=F)
system("liftOver data.file.bed /publicdata/liftover_20160308/hg18ToHg19.over.chain data.file.hg19.bed data.file.hg19.unlifted.bed")
df19<-read.table('data.file.hg19.bed', header=F)
df19<-merge(df19[,c(4,2)], data.file, by.x=1, by.y=2)
colnames(df19)[1:2]<-c("Proxy", "Coordinate_HG19")
df19<-df19[,-c(7,9)]
write.table(df19, "data.file.hg19.txt", row.names=F,col.names=T, sep="\t")
    return(df19)
    }

#### regional plot of GWAS P-values taken from summary statistics

gwas_plot <- function(data, snp, snp_ld , place="topleft"){

#data    = merge(fgwas, data, by=1)
chro    = as.numeric(substring(data$Chromosome[1], 4))
fgwas_s = subset(fgwas, chr == chro & pos >= min(data$Coordinate_HG19) & pos <= max(data$Coordinate_HG19))
data    = merge(fgwas_s, data, by=1, all.x=T)
#data    = data[order(data$pos),]

data$log10p = -log(data$P_VALUE,10)
    
offset = round(((maxp-minp)*0.04),0)    
min.pos<-minp+offset 
max.pos<-maxp-offset

data$group=cut(data$RSquared, breaks=c(0,0.2,0.4,0.6,0.8,1))

color_scheme = c("#26285a", "#8ec7e0", "#69b34b", "#eb9d3c", "#d62a31")
par (pin= c(6,1.5))

plot(data$pos, data$log10p,  xlim=c(min.pos, max.pos), pch=c(21,23)[(data$rsid %in% c(snp,snp_ld)) +1], 
     cex=c(1.7), bg=color_scheme[data$group], xaxt="n", las=1 ,  xlab = "",  ylab = "-log10(P)", cex.axis=1.5)

axis(3, las=1, cex.axis=1.5)
snp_lables=as.character(data$rsid)
snp_lables[!(snp_lables %in% c(snp,snp_ld))]<-""
text(data$pos, data$log10p,  snp_lables,cex=1.5, pos=c(1,2) )
legend(place, legend = c("[0,0.2]", "(0.2,0.4]", "(0.4,0.6]", "(0.6,0.8]", "(0.8,1.0]"), fill=color_scheme, cex=1.5,
      title= "LD")

par(new=TRUE)
data.RR.ordered <- data[order(data$pos),]
    
plot(1, 1, type = "n", xlim = c(min.pos, max.pos), axes = FALSE,  xlab = "", ylim = c(0,100))
    abline(h = 0)
    lines(data.RR.ordered$Coordinate_HG19 ,data.RR.ordered$RecombinationRate, col = "skyblue3", lwd=1.5)
    axis(4, las=1, cex.axis=1.5)


}


### Regional plots using fGWAS results

fgwas_plot <- function(data, snp, snp_ld ){

#data    = merge(fgwas, data, by=1)
chro    = as.numeric(substring(data$Chromosome[1], 4))
fgwas_s = subset(fgwas, chr == chro & pos >= min(data$Coordinate_HG19) & pos <= max(data$Coordinate_HG19))
data    = merge(fgwas_s, data, by=1, all.x=T)
    
data$log10p = data$PPA
    
offset = round(((maxp-minp)*0.04),0)    
min.pos<-minp+offset 
max.pos<-maxp-offset

par (pin= c(6,0.8))

plot(data$pos, data$log10p,  xlim=c(min.pos, max.pos), pch=c(21,23)[(data$rsid %in% c(snp,snp_ld)) +1], 
     cex=c(1.7), bg="gray", xaxt="n", las=1 ,  xlab = "",  cex.axis=1.5, ylim=c(0,1))

data    = data[order(data$pos),]
chu     = data[!duplicated(data$chunk), c("pos" ,"chunk")]

abline(v=chu$pos, col="gray", lwd=2, lty=2)
    
#axis(3, las=1, cex.axis=1.5)
snp_lables=as.character(data$rsid)
snp_lables[!(snp_lables %in% c(snp,snp_ld))]<-""
text(data$pos, data$log10p,  snp_lables,
     pos=c(1,2) , cex=1.5)
}


### HiC plots

suppressPackageStartupMessages(library(RColorBrewer))
suppressPackageStartupMessages(library(gplots))

upperHalfCircle = function(x, y, r, nsteps = 100, color = "red")
{  
   rs = seq(0, pi, len = nsteps)
   xc = x + r * cos(rs)
   yc = y + r * sin(rs)
   lines(xc, yc, col = color) 
}

plot_loops = function(loops, chrom, minp, maxp, h=0.5e6 ){

offset = round(((maxp-minp)*0.04),0)    

min.pos<-minp+offset 
max.pos<-maxp-offset

loops_subs = loops[(loops$V1 == as.character(chrom) & ((loops$V2 >= min.pos & loops$V3 <= max.pos)  |
                  (loops$V2 >= min.pos & loops$V2 <= max.pos) | (loops$V3 >= min.pos & loops$V3 <= max.pos))),]
par (pin= c(6,1.5))
plot(1, 1, type = "n", xlim = c(min.pos, max.pos), ylab = "", axes= F, ylim=c(h,0))

for (loo in 1:nrow (loops_subs)){
    dist = loops_subs[loo,3]-loops_subs[loo,2]
upperHalfCircle(loops_subs[loo,2]+(dist/2) ,0, dist/2)

}
    }

loops = read.table("/frazer01/home/paola/Family1070/private_output/HiC_loops_correlation/pgltools/CM.reference.bed", sep="\t", skip=1)

### GNB4

snp <- "rs7612445"
setwd("/frazer01/home/paola/Family1070/private_output/LD_plots/GNB4/")
minp<-179058380
maxp<-179224867

# minp<-179170000
# maxp<-179174000

snapfile = "plotdata/snap102962034781136667.data.tmp"
data = liftover(snapfile)
pdf(paste(snp,"fgwas.pdf", sep="."))
par (mfrow=c(3,1))
gwas_plot(data, snp, NA)
fgwas_plot(data, snp, NA)

chrom=data$Chromosome[1]
plot_loops(loops, chrom, minp, maxp, h=0.5e6)
dev.off()

### MYH11

snp <- "rs3915499"
snp_ld <- "rs8044595"
setwd("/frazer01/home/paola/Family1070/private_output/LD_plots/MYH11/")
minp <-15831093
maxp <-16388888

minp <-15880000
maxp <-15940000


snapfile = "plotdata/snap2187412560162046309.data.tmp"
data = liftover(snapfile)

pdf(paste(snp,"fgwas_zin.pdf", sep="."))
par (mfrow=c(3,1))
gwas_plot(data, snp, snp_ld, "topright")
fgwas_plot(data, snp, snp_ld)

chrom=data$Chromosome[1]
plot_loops(loops, chrom, minp, maxp, h=0.5e6)
dev.off()

### PPTC7

setwd("/frazer01/home/paola/Family1070/private_output/LD_plots/PPTC7/")
snp <- "rs11065706"
snp_ld <- "rs6606689"

minp<-110820657
maxp<-111307345


snapfile = "plotdata/snap891252560191868931.data.tmp"
data = liftover(snapfile)


pdf(paste(snp,"fgwas.pdf", sep="."))
par (mfrow=c(3,1))
gwas_plot(data, snp, snp_ld, "topright")
fgwas_plot(data, snp, snp_ld)

chrom=data$Chromosome[1]
plot_loops(loops, chrom, minp, maxp, h=0.5e6)
dev.off()

### MEF2C

setwd("/frazer01/home/paola/Family1070/private_output/LD_plots/MEF2C/")
snp <- "rs176107"

minp<-87992939
maxp<-90113022
minp<-89218200
maxp<-89425850
minp<-89000000
maxp<-89500000


snapfile = "plotdata/snap1243516671757886653.data.tmp"
data = liftover(snapfile)

pdf(paste(snp,"fgwas.zin.pdf", sep="."))
par (mfrow=c(3,1))
gwas_plot(data, snp, NA, "topleft")
fgwas_plot(data, snp, NA)

chrom=data$Chromosome[1]
plot_loops(loops, chrom, minp, maxp, h=2e6)
dev.off()

### SCN10A

snp <- "rs6801957"
setwd("/frazer01/home/paola/Family1070/private_output/LD_plots/SCN10A/")
minp <-38590401
maxp <-38846604

minp <-38745823
maxp <-38788922

snapfile = "plotdata/snap8186206431260808222.data.tmp"
data = liftover(snapfile)

snp_ld <- c("rs10428132", "rs6599255","rs6790396", "rs6795970", "rs6800541", "rs7433306")

pdf(paste(snp,"fgwas_zin.pdf", sep="."))
par (mfrow=c(3,1))
gwas_plot(data, snp, snp_ld, "topleft")
fgwas_plot(data, snp, snp_ld)

chrom=data$Chromosome[1]
plot_loops(loops, chrom, minp, maxp, h=2e6)
dev.off()
